chore: extract out s3_utils_layer
