+++ lambda/layers/response_utils_layer/pyproject.toml
@@ -3,6 +3,4 @@ name = "kubrick_response_utils_layer"
 version = "0.1.0"
 description = "AWS Lambda Layer for HTTP response utilities"
 requires-python = ">=3.13"
-dependencies = [
-    "aioboto3>=15.0.0",
-]
+dependencies = []
